fix: rimraf coverage directories in --mode migrate
