+++ package.json
@@ -99,6 +99,7 @@
 		"prettier-plugin-curly": "^0.1.3",
 		"prettier-plugin-packagejson": "^2.4.7",
 		"release-it": "^17.0.0",
+		"rimraf": "^5.0.5",
 		"sentences-per-line": "^0.2.1",
 		"should-semantic-release": "^0.2.1",
 		"tsup": "^8.0.1",
