+++ src/initialize/initializeWithOptions.ts
@@ -1,12 +1,11 @@
 import { withSpinner, withSpinners } from "../shared/cli/spinners.js";
-import { createCleanUpFilesCommands } from "../shared/createCleanUpFilesCommands.js";
 import { GitHubAndOptions } from "../shared/options/readOptions.js";
 import { addOwnerAsAllContributor } from "../steps/addOwnerAsAllContributor.js";
 import { clearChangelog } from "../steps/clearChangelog.js";
 import { initializeGitHubRepository } from "../steps/initializeGitHubRepository/index.js";
 import { removeSetupScripts } from "../steps/removeSetupScripts.js";
 import { resetGitTags } from "../steps/resetGitTags.js";
-import { runCommands } from "../steps/runCommands.js";
+import { runCleanup } from "../steps/runCleanup.js";
 import { uninstallPackages } from "../steps/uninstallPackages.js";
 import { updateAllContributorsTable } from "../steps/updateAllContributorsTable.js";
 import { updateLocalFiles } from "../steps/updateLocalFiles.js";
@@ -61,10 +60,5 @@ export async function initializeWithOptions({
 		);
 	}
 
-	await runCommands(
-		"Cleaning up files",
-		createCleanUpFilesCommands({
-			bin: !!options.bin,
-		}),
-	);
+	await runCleanup(options);
 }
