+++ src/migrate/migrateWithOptions.ts
@@ -1,12 +1,11 @@
 import { withSpinner, withSpinners } from "../shared/cli/spinners.js";
-import { createCleanUpFilesCommands } from "../shared/createCleanUpFilesCommands.js";
 import { GitHubAndOptions } from "../shared/options/readOptions.js";
 import { clearUnnecessaryFiles } from "../steps/clearUnnecessaryFiles.js";
 import { detectExistingContributors } from "../steps/detectExistingContributors.js";
 import { finalizeDependencies } from "../steps/finalizeDependencies.js";
 import { initializeGitHubRepository } from "../steps/initializeGitHubRepository/index.js";
 import { populateCSpellDictionary } from "../steps/populateCSpellDictionary.js";
-import { runCommands } from "../steps/runCommands.js";
+import { runCleanup } from "../steps/runCleanup.js";
 import { updateAllContributorsTable } from "../steps/updateAllContributorsTable.js";
 import { updateLocalFiles } from "../steps/updateLocalFiles.js";
 import { writeReadme } from "../steps/writeReadme/index.js";
@@ -66,11 +65,5 @@ export async function migrateWithOptions({
 		await withSpinner("Populating CSpell dictionary", populateCSpellDictionary);
 	}
 
-	await runCommands(
-		"Cleaning up files",
-		createCleanUpFilesCommands({
-			bin: !!options.bin,
-			dedupe: true,
-		}),
-	);
+	await runCleanup(options);
 }
